// src/types/editClass.ts
export interface ClassData {
    category: string;
    hashtag: string[];
    title: string;
    class_type: string;
    description: string;
    difficulty: string;
    min_people: number;
    quantity: number;
    total_time: number;
    location: string;
    detail_location: string;
    price: number;
    image: string[];
  }